comsigns/requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# MediaPipe
mediapipe>=0.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
onnx>=1.14.0
onnxruntime>=1.15.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

